+++ src/climate_data/generate/draws.py
@@ -4,7 +4,6 @@ from collections import defaultdict
 import click
 import numpy as np
 import tqdm
-import xarray as xr
 
 from climate_data import cli_options as clio
 from climate_data import constants as cdc
@@ -24,17 +23,15 @@ def compile_gcm_main(
     historical_paths = list(
         (cdata.raw_annual_results / "historical" / target_variable).glob("*.nc")
     )
-    scenario_paths = list(
-        (cdata.raw_annual_results / cmip6_experiment / target_variable).glob(
+    scenario_paths = [
+        p
+        for p in (cdata.raw_annual_results / cmip6_experiment / target_variable).glob(
             f"*{gcm_member}.nc"
         )
-    )
+        if p.stem.split("_")[0] not in cdc.HISTORY_YEARS
+    ]
     print("Opening datasets")
-    ds = (
-        xr.open_mfdataset(historical_paths + scenario_paths, combine="by_coords")
-        .sortby("year")
-        .compute()
-    )
+    ds = cdata.combine_raw_annual(historical_paths + scenario_paths)
 
     print("Saving compiled dataset")
     transform = TRANSFORM_MAP[target_variable]